Move each widget into its own file
