+++ src/motile_tracker/import_export/__init__.py
@@ -1,2 +1,2 @@
-from .import_external_tracks_dialog import ImportTracksDialog  # noqa
+from .menus.import_external_tracks_dialog import ImportTracksDialog  # noqa
 from .load_tracks import tracks_from_df  # noqa
