+++ src/motile_tracker/import_export/menus/scale_widget.py
@@ -1,0 +1,66 @@
+from qtpy.QtCore import Qt
+from qtpy.QtWidgets import (
+    QDoubleSpinBox,
+    QFormLayout,
+    QLabel,
+    QVBoxLayout,
+    QWidget,
+)
+
+
+class ScaleWidget(QWidget):
+    """QWidget for specifying pixel calibration"""
+
+    def __init__(self, incl_z=True):
+        super().__init__()
+
+        self.incl_z = incl_z
+
+        layout = QVBoxLayout()
+
+        # Spinboxes for scaling in x, y, and z (optional)
+        layout.addWidget(QLabel("Specify scaling"))
+        scale_form_layout = QFormLayout()
+        self.z_spin_box = self._scale_spin_box()
+        self.y_spin_box = self._scale_spin_box()
+        self.x_spin_box = self._scale_spin_box()
+
+        if self.incl_z:
+            scale_form_layout.addRow(QLabel("z"), self.z_spin_box)
+        scale_form_layout.addRow(QLabel("y"), self.y_spin_box)
+        scale_form_layout.addRow(QLabel("x"), self.x_spin_box)
+
+        layout.addLayout(scale_form_layout)
+        layout.setAlignment(Qt.AlignTop)
+
+        self.setLayout(layout)
+
+    def _scale_spin_box(self) -> QDoubleSpinBox:
+        """Return a QDoubleSpinBox for scaling values"""
+
+        spin_box = QDoubleSpinBox()
+        spin_box.setValue(1.0)
+        spin_box.setSingleStep(0.1)
+        spin_box.setMinimum(0)
+        spin_box.setDecimals(3)
+        return spin_box
+
+    def get_scale(self) -> list[float]:
+        """Return the scaling values in the spinboxes as a list of floats.
+        Since we currently require a dummy 1 value for the time dimension, add it here."""
+
+        if self.incl_z:
+            scale = [
+                1,
+                self.z_spin_box.value(),
+                self.y_spin_box.value(),
+                self.x_spin_box.value(),
+            ]
+        else:
+            scale = [
+                1,
+                self.y_spin_box.value(),
+                self.x_spin_box.value(),
+            ]
+
+        return scale
